+++ src/lib/output/themes/default/partials/comment.tsx
@@ -1,6 +1,6 @@
 import type { DefaultThemeRenderContext } from "../DefaultThemeRenderContext";
 import { JSX, Raw } from "../../../../utils";
-import type { Reflection } from "../../../../models";
+import { Reflection, ReflectionKind } from "../../../../models";
 import { camelToTitleCase } from "../../lib";
 
 export function comment({ markdown }: DefaultThemeRenderContext, props: Reflection) {
@@ -8,10 +8,14 @@ export function comment({ markdown }: DefaultThemeRenderContext, props: Reflecti
 
     // Note: Comment modifiers are handled in `renderFlags`
 
+    const tags = props.kindOf(ReflectionKind.SomeSignature)
+        ? props.comment.blockTags.filter((tag) => tag.tag !== "@returns")
+        : props.comment.blockTags;
+
     return (
         <div class="tsd-comment tsd-typography">
             <Raw html={markdown(props.comment.summary)} />
-            {props.comment.blockTags.map((item) => (
+            {tags.map((item) => (
                 <>
                     <h3>{camelToTitleCase(item.tag.substring(1))}</h3>
                     <Raw html={markdown(item.content)} />