+++ src/lib/output/themes/default/assets/bootstrap.ts
@@ -7,6 +7,7 @@ import { initTheme } from "./typedoc/Theme";
 
 addEventListener("load", () => {
     initSearch();
+    initCopyCode();
 
     registerComponent(Toggle, "a[data-toggle]");
     registerComponent(Accordion, ".tsd-index-accordion");
@@ -21,3 +22,25 @@ addEventListener("load", () => {
 
     Object.defineProperty(window, "app", { value: app });
 });
+
+function initCopyCode() {
+    document.querySelectorAll("pre > button").forEach((button) => {
+        let timeout: ReturnType<typeof setTimeout>;
+        button.addEventListener("click", () => {
+            if (button.previousElementSibling instanceof HTMLElement) {
+                navigator.clipboard.writeText(
+                    button.previousElementSibling.innerText.trim()
+                );
+            }
+            button.textContent = "Copied!";
+            button.classList.add("visible");
+            clearTimeout(timeout);
+            timeout = setTimeout(() => {
+                button.classList.remove("visible");
+                timeout = setTimeout(() => {
+                    button.textContent = "Copy";
+                }, 100);
+            }, 1000);
+        });
+    });
+}